Simplify metabox: use checkbox instead of radio buttons, add status indicator

+++ admin/class-wcl-metabox.php
@@ -30,105 +30,63 @@ class WCL_Metabox {
         // Add nonce for security
         wp_nonce_field('wcl_save_metabox', 'wcl_metabox_nonce');
 
-        $post_setting = get_post_meta($post->ID, '_wcl_enable_paywall', true);
+        $enabled = get_post_meta($post->ID, '_wcl_enable_paywall', true);
         $preview_percentage = get_post_meta($post->ID, '_wcl_preview_percentage', true);
         $default_mode = get_option('wcl_default_paywall_mode', 'disabled');
-
-        // Determine if paywall is currently active for this post
-        $is_paywall_active = ($post_setting === 'yes') || ($post_setting === '' && $default_mode === 'enabled');
 
         // Use global setting if not set
         if ($preview_percentage === '') {
             $preview_percentage = get_option('wcl_preview_percentage', 30);
         }
+
+        // Is paywall currently active?
+        $is_checked = ($enabled === 'yes');
         ?>
         <div class="wcl-metabox-content">
-            <?php if ($default_mode === 'enabled') : ?>
-                <!-- Default is ENABLED: show option to disable -->
-                <p class="wcl-default-notice">
-                    <em><?php _e('Paywall is enabled by default for all posts.', 'wp-content-locker'); ?></em>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="" <?php checked($post_setting, ''); ?> />
-                        <?php _e('Use default (enabled)', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="no" <?php checked($post_setting, 'no'); ?> />
-                        <?php _e('Disable paywall for this post', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="yes" <?php checked($post_setting, 'yes'); ?> />
-                        <?php _e('Enable paywall (explicit)', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-            <?php else : ?>
-                <!-- Default is DISABLED: show option to enable -->
-                <p class="wcl-default-notice">
-                    <em><?php _e('Paywall is disabled by default for all posts.', 'wp-content-locker'); ?></em>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="" <?php checked($post_setting, ''); ?> />
-                        <?php _e('Use default (disabled)', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="yes" <?php checked($post_setting, 'yes'); ?> />
-                        <?php _e('Enable paywall for this post', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-                <p>
-                    <label>
-                        <input type="radio" name="wcl_enable_paywall" value="no" <?php checked($post_setting, 'no'); ?> />
-                        <?php _e('Disable paywall (explicit)', 'wp-content-locker'); ?>
-                    </label>
-                </p>
-            <?php endif; ?>
+            <p class="wcl-default-notice" style="background:#f0f0f1;padding:8px 10px;border-left:3px solid #2271b1;margin:0 0 10px;">
+                <?php if ($default_mode === 'enabled') : ?>
+                    <em><?php _e('Global default: Enabled', 'wp-content-locker'); ?></em>
+                <?php else : ?>
+                    <em><?php _e('Global default: Disabled', 'wp-content-locker'); ?></em>
+                <?php endif; ?>
+            </p>
 
-            <hr style="margin: 15px 0;">
+            <p>
+                <label>
+                    <input type="checkbox"
+                           name="wcl_enable_paywall"
+                           id="wcl_enable_paywall"
+                           value="yes"
+                           <?php checked($is_checked, true); ?> />
+                    <strong><?php _e('Enable paywall for this post', 'wp-content-locker'); ?></strong>
+                </label>
+            </p>
 
-            <p class="wcl-preview-percentage-wrap" style="<?php echo !$is_paywall_active ? 'opacity: 0.5;' : ''; ?>">
+            <p>
                 <label for="wcl_preview_percentage">
                     <?php _e('Preview percentage:', 'wp-content-locker'); ?>
                 </label>
-                <input type="number" name="wcl_preview_percentage" id="wcl_preview_percentage"
+                <input type="number"
+                       name="wcl_preview_percentage"
+                       id="wcl_preview_percentage"
                        value="<?php echo esc_attr($preview_percentage); ?>"
                        min="10" max="90" style="width: 60px;" /> %
                 <br>
-                <span class="description"><?php _e('Leave empty to use global setting.', 'wp-content-locker'); ?></span>
+                <span class="description" style="color:#646970;font-size:12px;">
+                    <?php _e('How much content to show before paywall.', 'wp-content-locker'); ?>
+                </span>
             </p>
+
+            <?php if ($enabled) : ?>
+            <p style="color:green;">
+                <strong><?php _e('Current status: Paywall is ON', 'wp-content-locker'); ?></strong>
+            </p>
+            <?php else : ?>
+            <p style="color:#666;">
+                <?php _e('Current status: Paywall is OFF', 'wp-content-locker'); ?>
+            </p>
+            <?php endif; ?>
         </div>
-
-        <style>
-            .wcl-default-notice {
-                background: #f0f0f1;
-                padding: 8px 10px;
-                border-left: 3px solid #2271b1;
-                margin: 0 0 10px;
-            }
-        </style>
-
-        <script>
-        jQuery(document).ready(function($) {
-            $('input[name="wcl_enable_paywall"]').on('change', function() {
-                var value = $('input[name="wcl_enable_paywall"]:checked').val();
-                var defaultMode = '<?php echo esc_js($default_mode); ?>';
-                var isActive = (value === 'yes') || (value === '' && defaultMode === 'enabled');
-
-                if (isActive) {
-                    $('.wcl-preview-percentage-wrap').css('opacity', '1');
-                } else {
-                    $('.wcl-preview-percentage-wrap').css('opacity', '0.5');
-                }
-            });
-        });
-        </script>
         <?php
     }
 
@@ -137,8 +95,11 @@ class WCL_Metabox {
      */
     public function save_meta_box($post_id) {
         // Check nonce
-        if (!isset($_POST['wcl_metabox_nonce']) ||
-            !wp_verify_nonce($_POST['wcl_metabox_nonce'], 'wcl_save_metabox')) {
+        if (!isset($_POST['wcl_metabox_nonce'])) {
+            return;
+        }
+
+        if (!wp_verify_nonce($_POST['wcl_metabox_nonce'], 'wcl_save_metabox')) {
             return;
         }
 
@@ -152,17 +113,17 @@ class WCL_Metabox {
             return;
         }
 
-        // Save enable paywall
-        // Values: 'yes' = explicitly enabled, 'no' = explicitly disabled, '' = use default
-        if (isset($_POST['wcl_enable_paywall'])) {
-            $enable_paywall = sanitize_text_field($_POST['wcl_enable_paywall']);
-            if (in_array($enable_paywall, array('yes', 'no', ''))) {
-                if ($enable_paywall === '') {
-                    delete_post_meta($post_id, '_wcl_enable_paywall');
-                } else {
-                    update_post_meta($post_id, '_wcl_enable_paywall', $enable_paywall);
-                }
-            }
+        // Check post type
+        if (get_post_type($post_id) !== 'post') {
+            return;
+        }
+
+        // Save enable paywall - checkbox version
+        if (isset($_POST['wcl_enable_paywall']) && $_POST['wcl_enable_paywall'] === 'yes') {
+            update_post_meta($post_id, '_wcl_enable_paywall', 'yes');
+        } else {
+            // Checkbox not checked = delete or set to 'no'
+            delete_post_meta($post_id, '_wcl_enable_paywall');
         }
 
         // Save preview percentage
